Require pip>8.0.0
The outdated command requires pip 8.0.0 or greater. In particular, it
requires the find_all_candidates method of pip.utils.PackageFinder which
was not introduced until 8.0.0. There's probably another way around this,
for example by pulling in PackageFinder or significantly rewriting outdated.py,
but I prefer not to.

By requiring pip>8.0.0, we thus drop support for platforms that don't
suport it. For example, Python 3.2 requires pip<8, typically 7.1.2.